// apps/desktop/src/generated/TaskManagerProcessDetails.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TaskManagerProcessDetailField } from "./TaskManagerProcessDetailField";
import type { TaskManagerProcessIdentity } from "./TaskManagerProcessIdentity";
import type { TaskManagerProcessIo } from "./TaskManagerProcessIo";
import type { TaskManagerProcessMemoryDetails } from "./TaskManagerProcessMemoryDetails";
import type { TaskManagerProcessState } from "./TaskManagerProcessState";

export type TaskManagerProcessDetails = {
  identity: TaskManagerProcessIdentity;
  parent_pid: number;
  name: string;
  state: TaskManagerProcessState;
  uid?: number | null;
  nice: number;
  thread_count: number;
  command_line?: string | null;
  executable?: string | null;
  cgroups: Array<string>;
  memory: TaskManagerProcessMemoryDetails;
  io?: TaskManagerProcessIo | null;
  voluntary_context_switches?: number | null;
  involuntary_context_switches?: number | null;
  restricted_fields: Array<TaskManagerProcessDetailField>;
};